상품 좋아요 요청시 상품의 좋아요 수가 변경되도록 수정
